delete interested from post to consumer

+++ app/src/main/java/com/example/shutaffim/View/PostItem.kt
@@ -38,7 +38,7 @@ fun PostItem(post: Post, navController: NavController, screen: Screen, postsVM: 
     val size = 100.dp
 
     val sdf = SimpleDateFormat("dd/MM/yyyy", Locale.getDefault())
-    val date = Date(post.date.toLong())
+    val date = Date(post.date)
     val dateString = sdf.format(date)
 
     ElevatedCard(
